Split the oauth-server in several modules
