+++ resource-server/src/main/java/com/alonso/spring/oauth2/controller/UserController.java
