+++ runtime/src/main/resources/db/migration/h2/V5__insert_authorities.sql
