+++ runtime/src/main/resources/db/migration/h2/V6__insert_user.sql
