+++ security-server/src/main/java/com/alonso/spring/oauth2/config/server/AuthServerOAuth2Config.java
