+++ security-server/src/main/java/com/alonso/spring/oauth2/service/UserDetailsServiceImpl.java
@@ -1,0 +1,30 @@
+package com.alonso.spring.oauth2.service;
+
+import com.alonso.spring.oauth2.model.User;
+import com.alonso.spring.oauth2.repository.UserRepository;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Optional;
+
+@Service
+public class UserDetailsServiceImpl implements UserDetailsService {
+
+    private final UserRepository userRepository;
+
+    public UserDetailsServiceImpl(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
+    @Override
+    @Transactional(readOnly = true)
+    public UserDetails loadUserByUsername(String username) {
+        User user = userRepository.findByUsername(username);
+
+        return Optional.ofNullable(user)
+                       .orElseThrow(() -> new UsernameNotFoundException(username));
+    }
+}
