--- src/main/java/com/alonso/spring/oauth2/repository/UserRepository.java
@@ -1,17 +1,0 @@
-package com.alonso.spring.oauth2.repository;
-
-import com.alonso.spring.oauth2.model.security.User;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    @Query("SELECT DISTINCT user FROM User user " +
-            "INNER JOIN FETCH user.authorities AS authorities " +
-            "WHERE user.username = :username")
-    User findByUsername(@Param("username") String username);
-}
